C: remove leftover loop, seed random generator with time()
The loop would run until the 100th win, which would require several
billions of questions, but obviously there are not enough, so since the
question arrays are not circular, it stops with a segfault. That loop
must be some leftover debug aid, not present in other implementations,
so it's safe to be removed.